Added CSV Capabilities
Cleaned up code and added CSV capabilities. Some work still needs to be
done so that the program filters for Windows XP. As in the previous
version, it still prints the output of Nmap. This program is still not
to be used to find Windows XP computers on the network in its current
state.